Shows Taliesen West
I am new to Swift. I have been making some apps to practice. The goal is to add multiple annotations to the map, but I am trying to figure that out. So if anyone would like to add to this to show me how it is done that would be great!

Next location:
Fallingwater Latitude 39.906299 Longitude -79.467794